Massive rabbitmq rework
