+++ ares/docker/docker-compose-local.yml.j2
@@ -6,7 +6,7 @@ volumes:
     volume_{{ service_name }}_{{ team_id }}:
     {% endfor %}
 {% endfor %}
-    
+
 networks:
     ictf_net:
 
@@ -17,9 +17,9 @@ services:
         hostname: database
         ports:
             - "5000:80"
-        networks: 
+        networks:
           ictf_net:
-            aliases: 
+            aliases:
               - database.ictf
         sysctls:
           - net.core.somaxconn=65535
@@ -29,21 +29,22 @@ services:
     gamebot:
         image: ictf_gamebot
         hostname: gamebot
-        networks: 
+        networks:
           ictf_net:
-            aliases: 
+            aliases:
               - gamebot.ictf
         environment:
             - LOGSTASH_ID=gamebot
-        
+            - NUM_SCRIPTBOTS={{ num_scriptbots }}
+
     scoreboard:
         image: ictf_scoreboard
         hostname: scoreboard
-        ports: 
+        ports:
             - "80:80"
-        networks: 
+        networks:
           ictf_net:
-            aliases: 
+            aliases:
               - scoreboard.ictf
         environment:
             - LOGSTASH_ID=scoreboard
@@ -51,11 +52,11 @@ services:
     teaminterface:
         image: ictf_teaminterface
         hostname: teaminterface
-        ports: 
+        ports:
             - "8080:80"
-        networks: 
+        networks:
           ictf_net:
-            aliases: 
+            aliases:
               - teaminterface.ictf
         environment:
             - LOGSTASH_ID=teaminterface
@@ -63,15 +64,27 @@ services:
     logger:
         image: ictf_logger
         hostname: logger
-        ports: 
+        ports:
             - "5601:5601"
             - "3000:3000"
-        networks: 
+        networks:
           ictf_net:
-            aliases: 
+            aliases:
               - logger.ictf
         environment:
             - LOGSTASH_ID=logger
+
+    dispatcher:
+      image: ictf_dispatcher
+      hostname: dispatcher
+      ports:
+          - "5672:5672"
+          - "15672:15672"
+      networks:
+        ictf_net:
+          aliases:
+            - dispatcher.ictf
+
 
 #    router:
 #        image: ictf_router
@@ -88,8 +101,6 @@ services:
               target: /var/run/docker.sock
 
         environment:
-            - BOT_ID={{ scriptbot_id }}
-            - ALL_BOTS={{ num_scriptbots }}
             - IS_LOCAL_REGISTRY=1
             - API_SECRET={{ api_secret }}
             - REGISTRY_USERNAME="dummy"
@@ -97,9 +108,9 @@ services:
             - REGISTRY_ENDPOINT="dummy"
             - LOGSTASH_ID=scriptbot{{ scriptbot_id }}
 
-        networks: 
+        networks:
           ictf_net:
-            aliases: 
+            aliases:
               - scriptbot{{ scriptbot_id }}.ictf
 {% endfor %}
 {% for team_id in range(1, num_teams + 1) %}
@@ -114,10 +125,10 @@ services:
 
         environment:
             - TEAM_ID={{ team_id }}
-        
+
         entrypoint: /bin/bash -c "{{ teamvm_entrypoints[team_id - 1] }}"
 
-        networks: 
+        networks:
           ictf_net:
 
 {% endfor %}
@@ -131,7 +142,7 @@ services:
         volumes:
             - volume_{{ service_name }}_{{ team_id }}:/home/chall/service/
 
-        networks: 
+        networks:
           ictf_net:
     {% endfor %}
 {% endfor %}
