+++ gamebot/requirements.txt
@@ -3,3 +3,4 @@ coloredlogs==7.3.1
 requests==2.18.4
 setuptools==44.0.0
 python-logstash
+pika
