+++ gamebot/start.sh
