+++ hephaestus/1_dispatcher/Dockerfile
@@ -1,0 +1,1 @@
+FROM rabbitmq:management
