+++ hephaestus/1_scriptbot/Dockerfile
@@ -11,6 +11,6 @@ RUN pip3 install -r requirements.txt
 
 RUN ansible-playbook provisioning/hephaestus_provisioning/ansible-provisioning.yml
 
-RUN chmod +x ./start.sh 
+RUN chmod +x ./start.sh
 
 ENTRYPOINT ./start.sh
